requirements.txt
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.53.1
imageio==2.34.2
kiwisolver==1.4.5
lazy_loader==0.4
lxml==5.2.2
matplotlib==3.9.1.post1
networkx==3.3
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyimage==0.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
scikit-image==0.24.0
scipy==1.14.0
six==1.16.0
tifffile==2024.7.24
typing_extensions==4.12.2
tzdata==2024.1
